added chnages to improve code according to review
